scroll profile view
