+++ Ani/View/Profile/Contents/Clip/ANIProfileClipCell.swift
@@ -1,0 +1,24 @@
+//
+//  ANIProfileClipCell.swift
+//  Ani
+//
+//  Created by 전민섭 on 2018/04/19.
+//  Copyright © 2018年 JeonMinseop. All rights reserved.
+//
+
+import UIKit
+
+class ANIProfileClipCell: UITableViewCell {
+  override init(style: UITableViewCellStyle, reuseIdentifier: String?) {
+    super.init(style: style, reuseIdentifier: reuseIdentifier)
+    setup()
+  }
+  
+  required init?(coder aDecoder: NSCoder) {
+    fatalError("init(coder:) has not been implemented")
+  }
+  
+  private func setup() {
+    backgroundColor = .blue
+  }
+}
